Toan

+++ public/js/delete-user.js
@@ -1,21 +1,23 @@
 const deleteUser = async (event) => {
-    event.preventDefault()
-    console.log('Deleting user')
+  if (confirm("Are you sure you want to delete your account?")) {
+    event.preventDefault();
+    console.log("Deleting user");
 
     // add the controller route for the delete user by session id
-    const response = await fetch('/api/user/delete', {
-        method: "DELETE",
+    const response = await fetch("/api/user/delete", {
+      method: "DELETE",
     });
 
     if (response.ok) {
-        // takes back to homepage after the delete
-        document.location.replace('/')
+      // takes back to homepage after the delete
+      document.location.replace("/");
     } else {
-        alert('Error in deleting user')
+      alert("Error in deleting user");
     }
+  } else {
+    return;
+  }
 };
 
-// selects button by id and adds functionality 
-document
-    .querySelector('#delete-user')
-    .addEventListener('click', deleteUser)
+// selects button by id and adds functionality
+document.querySelector("#delete-user").addEventListener("click", deleteUser);
